+++ NEWS.md
@@ -1,3 +1,9 @@
+# scorecard 0.1.3
+
+* Modified the format of printing message.
+* Added the split_df function which split a dataframe into two.
+* Reorder the binning information. Put the missing as the first binning.
+
 # scorecard 0.1.2
 
 * fixed a bug in var_filter